A port to HP/UX. Thanks to Jim Foraker for the patch.  [ PATCH 1733645 ] Changes include: * no wait4(), so emulated with wait3(). Another option would be to just use wait3() on all platforms, since the extra arg of wait4() isn't currently being used. * check for strtoll() and use strtol if it doesn't exist, even if off_t is 64-bit. Some people have used __strtoll() on HP/UX, but in radmind it's only used for progress reporting, so it didn't seem worth it to use an unapproved function. * use = instead of == in aclocal, since HP/UX /bin/sh doesn't support that test * define _XOPEN_SOURCE and _XOPEN_SOURCE_EXTENDED=1 on HP/UX. This not only fixes a bunch of prototypes, but is necessary to get gcc to successfully parse the HP/UX includes. * #include <arpa/inet.h> a few places it probably should have been anyways, for hton* and ntoh*. threw in a cast in repo.c to fix a compiler warning probably occuring on multiple platforms.

+++ repo.c
@@ -9,6 +9,7 @@
 #include <sys/param.h>
 #include <sys/time.h>
 #include <netinet/in.h>
+#include <arpa/inet.h>
 #include <ctype.h>
 #include <netdb.h>
 #include <stdio.h>
@@ -134,7 +135,7 @@ main( int argc, char *argv[] )
 	    err++;
 	} else {
 	    for ( i = 0; i < len; i++ ) {
-		if ( isspace( event[ i ] )) {
+		if ( isspace( (int)event[ i ] )) {
 		    err++;
 		    break;
 		}
